完成供应商的列表功能，删除功能，详情功能，公司名查找功能
