+++ views/blueTpl/Header.html
@@ -14,7 +14,7 @@
     <h1>超市账单管理系统</h1>
     <div class="publicHeaderR">
         <p><span id="time">下午好！</span><span style="color: #fff21b"> Admin</span> , 欢迎你！</p>
-        <a href="login.html">退出</a>
+        <a href="/login">退出</a>
     </div>
 </header>
 <!--时间--    >
@@ -29,10 +29,10 @@
         <nav>
             <ul class="list">
                 <li><a href="billList.html">账单管理</a></li>
-                <li><a href="providerList.html">供应商管理</a></li>
+                <li><a href="/providerList">供应商管理</a></li>
                 <li  id="active"><a href="userList.html">用户管理</a></li>
-                <li><a href="password.html">密码修改</a></li>
-                <li><a href="login.html">退出系统</a></li>
+                <li><a href="/password">密码修改</a></li>
+                <li><a href="/login">退出系统</a></li>
             </ul>
         </nav>
     </div>
