+++ views/blueTpl/login.html
@@ -3,7 +3,7 @@
 <head lang="en">
     <meta charset="UTF-8">
     <title>系统登录 - 超市账单管理系统</title>
-    <link rel="stylesheet" href="css/style.css"/>
+    <link rel="stylesheet" href="/static/css/style.css"/>
 </head>
 <body class="login_bg">
     <section class="loginBox">
