+++ views/blueTpl/providerAdd.html
@@ -1,40 +1,4 @@
-﻿<!DOCTYPE html>
-<html>
-<head lang="en">
-    <meta charset="UTF-8">
-    <title>超市账单管理系统</title>
-    <link rel="stylesheet" href="css/public.css"/>
-    <link rel="stylesheet" href="css/style.css"/>
-</head>
-<body>
-<!--头部-->
-<header class="publicHeader">
-    <h1>超市账单管理系统</h1>
-
-    <div class="publicHeaderR">
-        <p><span>下午好！</span><span style="color: #fff21b"> Admin</span> , 欢迎你！</p>
-        <a href="login.html">退出</a>
-    </div>
-</header>
-<!--时间-->
-<section class="publicTime">
-    <span id="time">2015年1月1日 11:11  星期一</span>
-    <a href="#">温馨提示：为了能正常浏览，请使用高版本浏览器！（IE10+）</a>
-</section>
-<!--主体内容-->
-<section class="publicMian ">
-    <div class="left">
-        <h2 class="leftH2"><span class="span1"></span>功能列表 <span></span></h2>
-        <nav>
-            <ul class="list">
-                <li><a href="billList.html">账单管理</a></li>
-                <li id="active"><a href="providerList.html">供应商管理</a></li>
-                <li><a href="userList.html">用户管理</a></li>
-                <li><a href="password.html">密码修改</a></li>
-                <li><a href="login.html">退出系统</a></li>
-            </ul>
-        </nav>
-    </div>
+﻿{{ template "header"}}
     <div class="right">
         <div class="location">
             <strong>你现在所在的位置是:</strong>
@@ -81,17 +45,50 @@
                 <div class="providerAddBtn">
                     <!--<a href="#">保存</a>-->
                     <!--<a href="providerList.html">返回</a>-->
-                    <input type="button" value="保存" onclick="history.back(-1)"/>
+                    <input type="button" value="保存" class="saveBtn" />
                     <input type="button" value="返回" onclick="history.back(-1)"/>
                 </div>
             </form>
         </div>
-
     </div>
 </section>
 <footer class="footer">
 </footer>
-<script src="js/time.js"></script>
-
+<script src="/static/js/time.js"></script>
+<script>
+    $(function () {
+       $(".saveBtn").click(function () {
+           providerId = $("#providerId").val();
+           providerName = $("#providerName").val();
+           people = $("#people").val();
+           phone = $("#phone").val();
+           address = $("#address").val();
+           fax = $("#fax").val();
+           describe = $("#describe").val();
+           $.ajax({
+               type:"GET",
+               url:"/providerAddData",
+               data:{
+                   providerId:providerId,
+                   providerName:providerName,
+                   people:people,
+                   phone:phone,
+                   address:address,
+                   fax:fax,
+                   describe:describe
+               },
+               dataType:"json",
+               success: function (res) {
+                  if(res) {
+                      alert("添加成功！");
+                      location.href = "/providerList"
+                  } else {
+                      alert("添加失败！");
+                  }
+               }
+           });
+       });
+    });
+</script>
 </body>
 </html>
