+++ views/blueTpl/providerList.html
@@ -1,40 +1,4 @@
-﻿<!DOCTYPE html>
-<html>
-<head lang="en">
-    <meta charset="UTF-8">
-    <title>超市账单管理系统</title>
-    <link rel="stylesheet" href="css/public.css"/>
-    <link rel="stylesheet" href="css/style.css"/>
-</head>
-<body>
-<!--头部-->
-<header class="publicHeader">
-    <h1>超市账单管理系统</h1>
-
-    <div class="publicHeaderR">
-        <p><span>下午好！</span><span style="color: #fff21b"> Admin</span> , 欢迎你！</p>
-        <a href="login.html">退出</a>
-    </div>
-</header>
-<!--时间-->
-<section class="publicTime">
-    <span id="time">2015年1月1日 11:11  星期一</span>
-    <a href="#">温馨提示：为了能正常浏览，请使用高版本浏览器！（IE10+）</a>
-</section>
-<!--主体内容-->
-<section class="publicMian ">
-    <div class="left">
-        <h2 class="leftH2"><span class="span1"></span>功能列表 <span></span></h2>
-        <nav>
-            <ul class="list">
-                <li><a href="billList.html">账单管理</a></li>
-                <li id="active"><a href="providerList.html">供应商管理</a></li>
-                <li><a href="userList.html">用户管理</a></li>
-                <li><a href="password.html">密码修改</a></li>
-                <li><a href="login.html">退出系统</a></li>
-            </ul>
-        </nav>
-    </div>
+﻿{{ template "header" }}
     <div class="right">
         <div class="location">
             <strong>你现在所在的位置是:</strong>
@@ -42,9 +6,9 @@
         </div>
         <div class="search">
             <span>供应商名称：</span>
-            <input type="text" placeholder="请输入供应商的名称"/>
-            <input type="button" value="查询"/>
-            <a href="providerAdd.html">添加供应商</a>
+            <input type="text" class="providerName" placeholder="请输入供应商的名称"/>
+            <input type="button" class="searchProvider" value="查询"/>
+            <a href="/providerAdd">添加供应商</a>
         </div>
         <!--供应商操作表格-->
         <table class="providerTable" cellpadding="0" cellspacing="0">
@@ -57,34 +21,26 @@
                 <th width="10%">创建时间</th>
                 <th width="30%">操作</th>
             </tr>
-            <tr>
-                <td>PRO-CODE—001</td>
-                <td>测试供应商001</td>
-                <td>韩露</td>
-                <td>15918230478</td>
-                <td>15918230478</td>
-                <td>2015-11-12</td>
-                <td>
-                    <a href="providerView.html"><img src="img/read.png" alt="查看" title="查看"/></a>
-                    <a href="providerUpdate.html"><img src="img/xiugai.png" alt="修改" title="修改"/></a>
-                    <a href="#" class="removeProvider"><img src="img/schu.png" alt="删除" title="删除"/></a>
-                </td>
-            </tr>
-            <tr>
-                <td>PRO-CODE—001</td>
-                <td>测试供应商001</td>
-                <td>韩露</td>
-                <td>15918230478</td>
-                <td>15918230478</td>
-                <td>2015-11-12</td>
-                <td>
-                    <a href="providerView.html"><img src="img/read.png" alt="查看" title="查看"/></a>
-                    <a href="providerUpdate.html"><img src="img/xiugai.png" alt="修改" title="修改"/></a>
-                    <a href="#" class="removeProvider"><img src="img/schu.png" alt="删除" title="删除"/></a>
-                </td>
-            </tr>
+            {{ range $key,$val:= .provider }}
+                <tr>
+                    <td>{{ $val.ProviderNumber }}</td>
+                    <td>{{ $val.ProviderName }}</td>
+                    <td>{{ $val.People }}</td>
+                    <td>{{ $val.PhoneNumber }}</td>
+                    <td>{{ $val.Fax }}</td>
+                    <td>{{ $val.Createtime }}</td>
+                    <td>
+                        <a href="/providerView?id={{ $val.Id }}"><img src="/static/img/read.png" alt="查看" title="查看"/></a>
+                        <a href="/providerUpdate?id={{ $val.Id }}"><img src="/static/img/xiugai.png" alt="修改" title="修改"/></a>
+                        <a href="#" class="removeProvider" did="{{ $val.Id }}"><img src="/static/img/schu.png" alt="删除" title="删除"/></a>
+                    </td>
+               </tr>
+            {{ else }}
+                <tr aria-colspan="7">
+                    <td>没有数据！</td>
+                </tr>
+            {{ end }}
         </table>
-
     </div>
 </section>
 
@@ -95,7 +51,7 @@
        <h2>提示</h2>
        <div class="removeMain" >
            <p>你确定要删除该供应商吗？</p>
-           <a href="#" id="yes">确定</a>
+           <a href="#" id="yes" did="">确定</a>
            <a href="#" id="no">取消</a>
        </div>
    </div>
@@ -105,9 +61,38 @@
 <footer class="footer">
 </footer>
 
-<script src="js/jquery.js"></script>
-<script src="js/js.js"></script>
-<script src="js/time.js"></script>
-
+<script src="/static/js/js.js"></script>
+<script src="/static/js/time.js"></script>
+<script>
+    $(function () {
+       $(".searchProvider").click(function () {
+            providerName = $(".providerName").val();
+            location.href = '/providerList?providerName=' + providerName
+       });
+       $(".removeProvider").click(function () {
+           did = $(this).attr("did");
+           $("#yes").attr("did", did);
+       });
+       $("#yes").click(function () {
+           did = $(this).attr("did");
+           $.ajax({
+               type:"GET",
+               url:"/providerDel",
+               data: {
+                   id:did
+               },
+               dataType:"json",
+               success:function (res) {
+                  if(res) {
+                     alert("删除成功！");
+                     location.reload();
+                  } else {
+                      alert("删除失败！");
+                  }
+               }
+           });
+       });
+    });
+</script>
 </body>
 </html>
